- Added int, float, double explicit instantiations - Fixed template parameters

+++ src/main.cpp
@@ -5,7 +5,7 @@
 int main()
 {
     std::cout << "Starting app" << std::endl;
-    GAlg::GVector<int> vec(12);
+    GAlg::GVector<double> vec(12);
 
     vec[0] = 4; vec[1] = 3; vec[2] = 14; vec[3] = 3; vec[4] = 312; vec[5] = 21;
     vec[6] = 3; vec[7] = 56; vec[8] = 34; vec[9] = 67; vec[10] = 22; vec[11] = 44;
